// packages/core/src/Types/Const/SHEET_ACTION_NAMES.ts
/**
 * Name set of all action
 */
export enum SHEET_ACTION_NAMES {
    /** @see src/Command/Action/InsertDataRowAction */
    INSERT_ROW_DATA_ACTION = 'InsertRowDataAction',
    REMOVE_ROW_DATA_ACTION = 'RemoveRowDataAction',
    INSERT_ROW_ACTION = 'InsertRowAction',
    REMOVE_ROW_ACTION = 'RemoveRowAction',
    SET_RANGE_DATA_ACTION = 'SetRangeDataAction',
    DELETE_RANGE_ACTION = 'DeleteRangeAction',
    INSERT_SHEET_ACTION = 'InsertSheetAction',
    REMOVE_SHEET_ACTION = 'RemoveSheetAction',
    CLEAR_RANGE_ACTION = 'ClearRangeAction',
    INSERT_RANGE_ACTION = 'InsertRangeAction',
    SET_TAB_COLOR_ACTION = 'SetTabColorAction',
    HIDE_SHEET_ACTION = 'SetWorkSheetHideAction',
    SET_WORKSHEET_ACTIVATE_ACTION = 'SetWorkSheetActivateAction',
    INSERT_COLUMN_ACTION = 'InsertColumnAction',
    INSERT_COLUMN_DATA_ACTION = 'InsertColumnDataAction',
    REMOVE_COLUMN_ACTION = 'RemoveColumnAction',
    REMOVE_COLUMN_DATA_ACTION = 'RemoveColumnDataAction',
    SET_ROW_HIDE_ACTION = 'SetRowHideAction',
    SET_COLUMN_HIDE_ACTION = 'SetColumnHideAction',
    SET_ROW_SHOW_ACTION = 'SetRowShowAction',
    SET_COLUMN_SHOW_ACTION = 'SetColumnShowAction',
    ADD_MERGE_ACTION = 'AddMergeAction',
    REMOVE_MERGE_ACTION = 'RemoveMergeAction',
    SET_WORKSHEET_NAME_ACTION = 'SetWorkSheetNameAction',
    SET_WORKSHEET_STATUS_ACTION = 'SetWorkSheetStatusAction',
    SET_COLUMN_WIDTH_ACTION = 'SetColumnWidthAction',
    SET_ROW_HEIGHT_ACTION = 'SetRowHeightAction',
    SET_HIDDEN_GRIDLINES_ACTION = 'SetHiddenGridlinesAction',
    SET_RIGHT_TO_LEFT_ACTION = 'SetRightToLeftAction',
    DELETE_BANDING_ACTION = 'DeleteBandingAction',
    SET_BANDING_ACTION = 'SetBandingAction',
    ADD_BANDING_ACTION = 'AddBandingAction',
    SET_SHEET_ORDER_ACTION = 'SetSheetOrderAction',
    SET_ZOOM_RATIO_ACTION = 'SetZoomRatioAction',
    SET_FROZEN_COLUMNS_ACTION = 'SetFrozenColumnsAction',
    SET_FROZEN_ROWS_ACTION = 'SetFrozenRowsAction',
}